+++ .github/CODEOWNERS
@@ -6,3 +6,4 @@ modules/vote @ProgramPhoenix
 lib @c0derMo
 lib/tasks @Kr0nox
 index.ts @c0derMo
+tests/discord.js-mock @c0derMo
